代码规范告警修改

+++ services/engine/codec/audio/decoder/audio_opus_decoder_plugin.cpp
@@ -136,7 +136,7 @@ int32_t AudioOpusDecoderPlugin::ProcessSendData(const std::shared_ptr<AudioBuffe
         bool isEos = inputBuffer->CheckIsEos();
         len = attr.size;
 
-        if ((len <= 0 || len > inputBuffer->GetBufferSize()) && !isEos) {
+        if ((len <= 0 || (uint32_t) len > inputBuffer->GetBufferSize()) && !isEos) {
             AVCODEC_LOGE("SendBuffer error buffer size:%{public}d", len);
             return AVCodecServiceErrCode::AVCS_ERR_INVALID_DATA;
         }
